Revert "Introduce wisp-client (#1826)"
This reverts commit cd2ff3911c790a9513d3a181ab952c996b102684 for breaking downstream
consumers due to injection failure:

```
    java.lang.NoClassDefFoundError: misk/client/EnvoyClientEndpointProvider
```
