+++ misk/build.gradle
@@ -60,7 +60,6 @@ dependencies {
   implementation project(':misk-metrics')
   implementation project(':misk-prometheus')
   implementation project(':misk-service')
-  api project(':wisp-client')
   api project(':wisp-deployment')
 
   testImplementation dep.kotlinxCoroutines
