+++ misk/src/main/java/misk/client/TunnelingUnixSocket.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package wisp.client;
+package misk.client;
 
 import java.io.File;
 import java.io.IOException;
