+++ misk/src/main/kotlin/misk/client/HttpClientConfigUrlProvider.kt
@@ -1,7 +1,6 @@
 package misk.client
 
 import com.google.inject.Inject
-import wisp.client.EnvoyClientEndpointProvider
 
 /**
  * Calculates the url for an http client config,
@@ -11,13 +10,13 @@ import wisp.client.EnvoyClientEndpointProvider
 class HttpClientConfigUrlProvider @Inject constructor() {
   @Inject(optional = true) lateinit var envoyClientEndpointProvider: EnvoyClientEndpointProvider
 
-  fun getUrl(endpointConfig: HttpClientEndpointConfig): String = when {
-    endpointConfig.url != null -> endpointConfig.url
-    endpointConfig.envoy != null -> envoyClientEndpointProvider.url(
-      endpointConfig.envoy.toWispConfig()
-    )
-    else -> throw IllegalArgumentException(
-      "One of url or envoy configuration must be set for clients"
-    )
+  fun getUrl(endpointConfig: HttpClientEndpointConfig): String {
+    when {
+      endpointConfig.url != null -> return endpointConfig.url
+      endpointConfig.envoy != null -> return envoyClientEndpointProvider.url(endpointConfig.envoy)
+      else -> throw IllegalArgumentException(
+        "One of url or envoy configuration must be set for clients"
+      )
+    }
   }
 }
