+++ misk/src/main/kotlin/misk/client/HttpClientFactory.kt
@@ -1,13 +1,16 @@
 package misk.client
 
-import okhttp3.Interceptor
-import okhttp3.OkHttpClient
-import wisp.client.EnvoyClientEndpointProvider
 import misk.security.ssl.SslContextFactory
 import misk.security.ssl.SslLoader
+import okhttp3.Interceptor
+import okhttp3.OkHttpClient
+import okhttp3.Protocol
+import java.io.File
+import java.net.Proxy
 import javax.inject.Inject
 import javax.inject.Provider
 import javax.inject.Singleton
+import javax.net.ssl.X509TrustManager
 
 @Singleton
 class HttpClientFactory @Inject constructor(
@@ -16,22 +19,63 @@ class HttpClientFactory @Inject constructor(
   private val okHttpClientCommonConfigurator: OkHttpClientCommonConfigurator
 ) {
   @com.google.inject.Inject(optional = true)
-  var envoyClientEndpointProvider: EnvoyClientEndpointProvider? = null
+  lateinit var envoyClientEndpointProvider: EnvoyClientEndpointProvider
 
   @com.google.inject.Inject(optional = true)
   var okhttpInterceptors: Provider<List<Interceptor>>? = null
 
   /** Returns a client initialized based on `config`. */
   fun create(config: HttpClientEndpointConfig): OkHttpClient {
-    val delegate = wisp.client.HttpClientFactory(
-      sslLoader.delegate,
-      sslContextFactory.delegate,
-      okHttpClientCommonConfigurator.delegate,
-      envoyClientEndpointProvider,
-      okhttpInterceptors?.get()
-    )
+    // TODO(mmihic): Cache, proxy, etc
+    val builder = unconfiguredClient.newBuilder()
+    builder.retryOnConnectionFailure(false)
+    okHttpClientCommonConfigurator.configure(builder = builder, config = config)
+    config.clientConfig.ssl?.let {
+      val trustStore = sslLoader.loadTrustStore(it.trust_store)!!
+      val trustManagers = sslContextFactory.loadTrustManagers(trustStore.keyStore)
+      val x509TrustManager = trustManagers.mapNotNull { it as? X509TrustManager }.firstOrNull()
+        ?: throw IllegalStateException("no x509 trust manager in ${it.trust_store}")
+      val sslContext = sslContextFactory.create(it.cert_store, it.trust_store)
+      builder.sslSocketFactory(sslContext.socketFactory, x509TrustManager)
+    }
 
-    return delegate.create(config.toWispConfig())
+    require(config.envoy == null || config.clientConfig.unixSocketFile == null) {
+      "Setting both `envoy` and `unixSocketFile` on `HttpClientEndpointConfig` is not supported!"
+    }
+
+    require(config.envoy == null || config.clientConfig.protocols == null) {
+      "Setting both `envoy` and `protocols` on `HttpClientEndpointConfig` is not supported!"
+    }
+
+    config.clientConfig.unixSocketFile?.let {
+      builder.socketFactory(UnixDomainSocketFactory(File(it)))
+      // No DNS lookup needed since we're just sending the request over a socket.
+      builder.dns(NoOpDns)
+      // Proxy config not supported
+      builder.proxy(Proxy.NO_PROXY)
+    }
+
+    config.clientConfig.protocols
+      ?.map { Protocol.get(it) }
+      ?.let { builder.protocols(it) }
+
+    config.envoy?.let {
+      builder.socketFactory(
+        UnixDomainSocketFactory(envoyClientEndpointProvider.unixSocket(config.envoy))
+      )
+      // No DNS lookup needed since we're just sending the request over a socket.
+      builder.dns(NoOpDns)
+      // Proxy config not supported
+      builder.proxy(Proxy.NO_PROXY)
+      // OkHttp <=> envoy over h2 has bad interactions, and benefit is marginal
+      builder.protocols(listOf(Protocol.HTTP_1_1))
+    }
+
+    okhttpInterceptors?.let {
+      builder.interceptors().addAll(it.get())
+    }
+
+    return builder.build()
   }
 
   companion object {
