+++ misk/src/main/kotlin/misk/client/HttpClientModule.kt
@@ -4,9 +4,9 @@ import com.google.inject.Key
 import com.google.inject.Provider
 import com.squareup.moshi.Moshi
 import misk.inject.KAbstractModule
-import misk.inject.asSingleton
 import okhttp3.OkHttpClient
 import javax.inject.Inject
+import javax.inject.Singleton
 
 /** Provides an [OkHttpClient] and [ProtoMessageHttpClient] for a peer service */
 class HttpClientModule constructor(
@@ -22,10 +22,10 @@ class HttpClientModule constructor(
       else Key.get(ProtoMessageHttpClient::class.java, annotation)
     bind(httpClientKey)
       .toProvider(HttpClientProvider(name))
-      .asSingleton()
+      .`in`(Singleton::class.java)
     bind(protoMessageHttpClientKey)
       .toProvider(ProtoMessageHttpClientProvider(name, getProvider(httpClientKey)))
-      .asSingleton()
+      .`in`(Singleton::class.java)
   }
 
   private class HttpClientProvider(private val name: String) : Provider<OkHttpClient> {
