+++ settings.gradle
@@ -39,7 +39,6 @@ include ':misk-zookeeper'
 include ':misk-zookeeper-testing'
 include ':samples:exemplar'
 include ':samples:exemplarchat'
-include ':wisp-client'
 include ':wisp-deployment'
 include ':wisp-logging'
 include ':wisp-logging-testing'
