--- wisp-client/build.gradle
@@ -1,24 +1,0 @@
-dependencies {
-  api project(':wisp-ssl')
-  api project(':wisp-logging')
-  implementation dep.jnrUnixsocket
-  implementation dep.kotlinStdLibJdk8
-  implementation dep.moshiKotlin
-  implementation dep.okHttp
-  implementation dep.okio
-  implementation dep.openTracing
-  implementation dep.openTracingOkHttp
-  implementation dep.retrofit
-  implementation dep.retrofitMoshi
-  implementation dep.retrofitProtobuf
-  implementation dep.retrofitWire
-}
-
-afterEvaluate { project ->
-  project.tasks.dokka {
-    outputDirectory = "$rootDir/docs/0.x"
-    outputFormat = 'gfm'
-  }
-}
-
-apply from: "$rootDir/gradle-mvn-publish.gradle"
