--- wisp-client/gradle.properties
@@ -1,4 +1,0 @@
-POM_ARTIFACT_ID=misk-client
-POM_NAME=misk-client
-POM_DESCRIPTION=HTTP client for misk
-POM_PACKAGING=jar
