--- wisp-client/src/main/kotlin/wisp/client/NoOpDns.kt
@@ -1,10 +1,0 @@
-package wisp.client
-
-import okhttp3.Dns
-import java.net.InetAddress
-
-object NoOpDns : Dns {
-  private val loopback = listOf(InetAddress.getLoopbackAddress())
-
-  override fun lookup(hostname: String) = loopback
-}
